Show render blocking (Chrome 92) and largest contentful paint.
In the coming Chrome 92 the render blocking info will be fixed and
potetntially work as it should. That means we can highlight which assets
that are render blocking (and potentially render blocking).

We can also highlight the image (if it is an image) that is the
largest contentful paint object.

I prepared HARs generated from Chrome/Edge in Browsertime to include
this info and plan to release it when Chrome 92 is released.